==> analyses/90_plots_and_comparisons/97_response_to_icb.Rmd <==
---
output: html_document
params:
  cpus: 1
  artifact_dir: "/data/scratch/sturm/tmp"
  meta: NULL
  ici_tpm: "../../data/14_ici_treatment/Genentech_for_scissor/genentech.rds"
  ici_meta: "../../data/14_ici_treatment/Genentech_for_scissor/genentech_clinical_data.tsv"
  neutro_sigs: "../../data/30_downstream_analyses/neutrophils/neutrophil_analysis/artifacts/neutro_sigs.csv"
  cell_type_major_sigs: "/home/sturm/Downloads/cell_type_major_signatures.csv"
  input_dir: NULL
---

```{r}
library(conflicted)
library(dplyr)
conflict_prefer("filter", "dplyr")
conflict_prefer("select", "dplyr")
conflict_prefer("rename", "dplyr")
library(survminer)
library(survival)
library(tidyr)
library(readr)
library(stringr)
library(tibble)
library(ggbeeswarm)
library(cowplot)
library(readxl)
```

# Load data

## ICI

```{r}
ici_tpm = read_rds(params$ici_tpm)
ici_meta = read_tsv(params$ici_meta) |> rename(dataset=study)
```

## Signatures 

```{r}
neutro_sig_tab = read_csv(params$neutro_sigs)
neutro_sig = sapply(unique(neutro_sig_tab$signature), function(sig) { neutro_sig_tab |> filter(signature == !!sig) |> pull(gene_symbol)}, USE.NAMES = TRUE, simplify = FALSE)
cell_type_sig_tab = read_csv(params$cell_type_major_sigs)
cell_type_sig = sapply(unique(cell_type_sig_tab$signature), function(sig) { cell_type_sig_tab |> filter(signature == !!sig) |> pull(gene_symbol)}, USE.NAMES = TRUE, simplify = FALSE)
```

```{r}
all_sigs = c(neutro_sig, cell_type_sig)
neutro_sig_names = names(neutro_sig)
ct_sig_names = names(cell_type_sig)
sig_names = names(all_sigs)
```

# Functions to compute signature scores

## "Mean z score" scoring function
This is similar to the LCAM score, just there is no differentiation between markers for different cell-types. 
```{r}
#' Score signature genes as described in the Merad paper
#' 
#' This is a simple mean of the z-scored gene expression
get_lcam_scores = function(mixture, genes) {
  exclude_genes = setdiff(genes, rownames(mixture))
  if(length(exclude_genes)) {
    warning(sprintf("Excluded the following genes which are not in the mixture matrix: %s", paste(exclude_genes, collapse=", ")))
    genes = setdiff(genes, exclude_genes)
  }
  signature_expression = t(scale(t(log1p(mixture[genes, ,drop=FALSE]))))
  apply(signature_expression, MARGIN = 2, mean, na.rm=TRUE)
}

scoring_function = get_lcam_scores 
```

## Significance codes
```{r}
sig_codes = Vectorize(function(p) {
  f = \(x) factor(x, levels=c("p < 0.001", "p < 0.01", "p < 0.05", "p < 0.1", "p >= 0.1"))
  if(p < 0.001) {
    f("p < 0.001")
  } else if(p < 0.01) {
    f("p < 0.01")
  } else if (p < 0.05) {
    f("p < 0.05")
  } else if (p < 0.1) {
    f("p < 0.1")
  } else {
    f("p >= 0.1")
  }
})
```


# Prepare data
## ICI treatment data

```{r, fig.width=12, fig.height=7}
ici_scores = data.frame(sapply(names(all_sigs), function(sig) {
  scoring_function(ici_tpm, all_sigs[[sig]])
})) |> mutate(
    sig_trn_mean=(sig_tan+sig_nan)/2
) |> 
  as_tibble(rownames="sample")

ici_clin_and_scores = ici_meta |> 
  rename(sample=sample_id) |>
  inner_join(ici_scores) |>
  mutate(response_to_any = ifelse(study_arm == "Docetaxel", response_to_chemotherapy, response_to_ici), .before="response_to_chemotherapy") |>
  mutate_at(vars(type, dataset, study_arm, response_to_chemotherapy, response_to_ici, response_to_any), factor)

```

# LUAD vs LUSC

```{r, fig.width=7, fig.height=5}
ici_clin_and_scores |> select(type, !!neutro_sig_names) |>
  filter(!is.na(type)) |> 
  pivot_longer(-type, names_to="signature", values_to="score") |>
  ggplot(aes(x=signature, y=score, color=type)) + 
  theme_cowplot() + 
  background_grid() + 
  geom_boxplot() + 
  theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1))
```

```{r}
# ICI
lapply(neutro_sig_names, function(sig) {
  res = summary(lm(as.formula(sprintf("`%s` ~ type + dataset + gender + study_arm", sig)), data=ici_clin_and_scores))
  res$coefficients |> as_tibble(rownames="var") |> mutate(sig=sig, .before=1)
}) |> bind_rows() |> filter(var == "typeLUSC")
```

## Survival analysis

```{r, fig.width=6, fig.height=6}
tmp_sig_names = make.names(sig_names)
coxph_results = lapply(c("any", "Docetaxel", "MPDL3280A"), function(arm) {
    lapply(c("all", "LUSC", "LUAD"), function(type) {
        filter_type = if(type == "all") {c("LUAD", "LUSC") } else {type}
        filter_arm = if(arm == "any") {c("Docetaxel", "MPDL3280A")} else {arm}
        type_covar = ifelse(type == "all", "+ type ", "")
        arm_covar = ifelse(arm == "any", "+ study_arm", "")
        coxph_res = lapply(tmp_sig_names, function(col) {
           res = summary(coxph(as.formula(sprintf("Surv(time, status) ~ `%s` %s %s + dataset", col, type_covar, arm_covar)), data=ici_clin_and_scores |> 
                                 filter(type %in% filter_type, study_arm %in% filter_arm)))
           res$coefficients |> 
             as_tibble(rownames="var") |> 
             filter(grepl( col, var)) |> 
             mutate(var=col, type=type)
           
        }) |> bind_rows() |> mutate(sig_code = sig_codes(`Pr(>|z|)`)) |> 
          # arrange(-coef) |> 
          mutate(var=factor(var, levels=unique(var)))
        
        # print(coxph_res |> ggplot(aes(x=var, y=coef)) +
        #         geom_bar(aes(fill=sig_code), stat="identity", color="black") + 
        #         scale_fill_brewer(palette= "Blues", direction=-1) + 
        #         theme_cowplot()+ 
        #         theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1)) + 
        #         geom_hline(yintercept = 0, color="black") +
        #         facet_wrap(~type, ncol=1)+ 
        #         ggtitle(sprintf("Study arm: %s", arm)))
        
        coxph_res |> mutate(type=type)
    }) |> bind_rows() |> mutate(study_arm=arm)
}) |> bind_rows()

```

```{r, fig.width=4, fig.height=4}
lapply(c("all", "LUAD", "LUSC"), function(tumor_type) {
  tmp_df = coxph_results |> 
    filter(study_arm == "MPDL3280A", type==tumor_type) |> 
    filter(var %in% make.names(ct_sig_names)) |> 
    mutate(fdr = p.adjust(`Pr(>|z|)`, method="fdr")) |>
    mutate(var = str_replace_all(var, "\\.", " "))
  
  p = tmp_df |> mutate(var = factor(var, levels= arrange(tmp_df, -coef) |> pull("var"))) |> 
    filter(fdr < 0.1) |> 
    ggplot(aes(x=var, y=coef, fill=coef)) +
    geom_bar(stat='identity') +
    scale_fill_distiller(palette="RdBu") + 
    theme_cowplot() + 
    theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1)) + 
    background_grid() +
    geom_hline(yintercept=0) + 
    ylab("log(hazard ratio)") + xlab("cell type") + theme(legend.position = "none") + 
    ggtitle(tumor_type)

  ggsave(file.path(params$artifact_dir, sprintf("cell_type_sig_score_%s.pdf", tumor_type)), width=4, height=4)
  
  p
})
```

```{r, fig.width=4, fig.height=4}
plot_kaplan_meyer = function(clinical_data, column, q=c(0.25, .75), pvalue=TRUE) {
  tmp_col = clinical_data[[column]]
  cutoffs = quantile(tmp_col, q)
  group = rep(NA, length(tmp_col))
  group[tmp_col < cutoffs[1]] = "bottom 25%"
  group[tmp_col > cutoffs[2]] = "top 25%"
  clinical_data$group = group
  clinical_data = clinical_data |> filter(time < 365 * 5)
  fit = survfit(Surv(time, status) ~ group, data=clinical_data)
  ggsurvplot(fit, data=clinical_data, pval=pvalue, risk.table = FALSE, xlab="time in days")
}

lapply(c("any", "Docetaxel", "MPDL3280A"), function(arm) {
    lapply(c("LUAD", "LUSC", "all"), function(type) {
        filter_type = if(type == "all") {c("LUAD", "LUSC") } else {type}
        filter_arm = if(arm == "any") {c("Docetaxel", "MPDL3280A")} else {arm}
        lapply(c("sig_tan", "sig_nan", "sig_trn"), function(sig) {
          p = plot_kaplan_meyer(
            ici_clin_and_scores |> filter(type %in% filter_type, study_arm %in% filter_arm),
            sig,
            pvalue=sprintf("p=%*.3f, CoxPH reg.", 3, coxph_results |> filter(var == sig, type == !!type, study_arm == arm) |> pull(`Pr(>|z|)`)),
          ) + 
            ggtitle(sprintf("%s (%s, %s)", sig, type, arm))
          ggsave(file.path(params$artifact_dir, sprintf("KM_%s_%s_%s.pdf", sig, type, arm)), width=4.5, height=4)
          p
        })
    })
})
```

```{r}
summary(coxph(as.formula("Surv(time, status) ~ dataset + sig_trn * type * study_arm"), data = ici_clin_and_scores))
summary(coxph(as.formula("Surv(time, status) ~ dataset + sig_tan * type * study_arm"), data = ici_clin_and_scores))
summary(coxph(as.formula("Surv(time, status) ~ dataset + sig_nan * type * study_arm"), data = ici_clin_and_scores))
```

## Responder vs. Non-responder

### Response to ICI
```{r, fig.width=6, fig.height=6}
tmp_sig_names = neutro_sig_names 
ici_data = ici_clin_and_scores
glm_res = lapply(c("any", "Docetaxel", "MPDL3280A"), function(arm) {
  lapply(c("all", "LUSC", "LUAD"), function(type) {
      filter_arm = if(arm == "any") {c("Docetaxel", "MPDL3280A")} else {arm}
      filter_type = if(type == "all") {c("LUAD", "LUSC") } else {type}
      type_covar = ifelse(type == "all", "+ type ", "")
      arm_covar = ifelse(arm == "any", "+ study_arm", "")
      tmp_res = lapply(tmp_sig_names, function(col) {
        res = summary(glm(as.formula(sprintf("response_to_any ~ `%s` %s %s + dataset", col, type_covar, arm_covar)), family="binomial", data=ici_data |> filter(type %in% filter_type, study_arm %in% filter_arm)))
        res$coefficients |> as_tibble(rownames="var") |> filter(grepl( col, var)) |> mutate(var=col)
      }) |> 
        bind_rows() |> 
        mutate(sig_code = sig_codes(`Pr(>|z|)`)) |>
        # arrange(-Estimate) |> 
        mutate(var=factor(var, levels=unique(var)))
      
      tmp_res = tmp_res |> mutate(type=type, arm=arm)
      # print(tmp_res |> ggplot(aes(x=var, y=Estimate)) + 
      #         geom_bar(aes(fill=sig_code), stat="identity") +
      #         scale_fill_brewer(palette= "Blues", direction=-1)  + 
      #         theme_cowplot()+
      #         theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1)) +
      #         geom_hline(yintercept = 0, color="black") + 
      #         ggtitle(sprintf("Study arm: %s", arm)) +
      #         facet_wrap(~type, ncol=1))
      
      tmp_res
      
  }) |> bind_rows()
}) |> bind_rows()
```

```{r, fig.width=7, fig.height=5}
pvalue_df = glm_res |> filter(type=="all", arm=="MPDL3280A") |> 
  select(signature=var, pvalue=`Pr(>|z|)`) |> 
  filter(signature %in% c("sig_tan", "sig_nan", "sig_trn")) |> 
  mutate(pvalue = sprintf("p=%*.3f", 3, pvalue)) |>
  distinct()

p = ici_data |> select(type, response=response_to_ici, dataset, starts_with("sig_")) |>
    pivot_longer(-c(type, response, dataset), names_to="signature", values_to="score") |>
    filter(!is.na(response)) |> 
    inner_join(pvalue_df) |> 
    mutate(signature = factor(signature, levels=c("sig_tan", "sig_nan", "sig_trn"))) |> 
    mutate(response = recode(response, "0"="Non-responder", "1"="Responder")) |> 
    ggplot(aes(x=response, y=score)) + 
    geom_quasirandom(aes(color=type, shape=dataset)) +
    geom_boxplot(outlier.alpha=0, width=.2) + 
    geom_text(aes(label = pvalue), x=-Inf, y=Inf, data=pvalue_df, hjust = -0.2, vjust=1) +
    theme_cowplot() + 
    background_grid() + 
    facet_wrap(~signature, ncol = 7, scales="free_y") + 
    ggtitle(sprintf("Scoring function: %s", params$scoring_function)) + 
    theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1))

ggsave(file.path(params$artifact_dir, "ici_treatment_response.pdf"), width = 7, height=5)
p
```
