==> analyses/90_plots_and_comparisons/99b_cpdb_circosplot.Rmd <==
---
title: "CPDB interaction plots"
output: html_document
params:
  artifact_dir: "/home/sturm/Downloads"
  input_dir: "../../data/30_downstream_analyses/plots_and_comparisons/99_cpdb_analysis/artifacts"
  cpus: 1
  meta: NULL
---

```{r, include=FALSE}
library(conflicted)

library(ggplot2)
library(ggraph)
library(dplyr)
conflict_prefer("filter", "dplyr")
conflict_prefer("rename", "dplyr")
library(readr)
library(ggrepel)
library(igraph)
library(readxl)
```

```{r, include=FALSE}

COLORS = list(cell_type_coarse=list(
        "B cell"="#1f77b4",
        "Endothelial cell"="#ff7f0e",
        "Epithelial cell"="#279e68",
        "Macrophage/Monocyte"="#d62728",
        "Mast cell"="#aa40fc",
        "NK cell"="#8c564b",
        "Neutrophils"="#e377c2",
        "Plasma cell"="#b5bd61",
        "Stromal"="#17becf",
        "T cell"="#aec7e8",
        "cDC"="#ffbb78",
        "pDC"="#98df8a",
        "Tumor cells"="#999999"
))

zeileis_28 = c(
    "#023fa5",
    "#7d87b9",
    "#bec1d4",
    "#d6bcc0",
    "#bb7784",
    "#8e063b",
    "#4a6fe3",
    "#8595e1",
    "#b5bbe3",
    "#e6afb9",
    "#e07b91",
    "#d33f6a",
    "#11c638",
    "#8dd593",
    "#c6dec7",
    "#ead3c6",
    "#f0b98d",
    "#ef9708",
    "#0fcfc0",
    "#9cded6",
    "#d5eae7",
    "#f3e1eb",
    "#f6c4e1",
    "#f79cd4",
    # these last ones were added:
    '#7f7f7f',
    "#c7c7c7",
    "#1CE6FF",
    "#336600"
)


immune_cells = c(
  "B cell"   ,
  "cDC",
  "Macrophage/Monocyte",
  "Mast cell",
  "Neutrophils",
  "NK cell",
  "pDC",
  "Plasma cell",
  "T cell"
)
```


```{r, include=FALSE}
cpdb_plot = function(
  cpdb_res,
  edge_color="log2_fc",
  edge_color_scale=scale_edge_color_gradientn(colors=pals::coolwarm(100), limits=c(-1,1) * max(abs(conn_meta[[edge_color]]))),
  cell_type_color_scale=zeileis_28
) {
  # make edges df
  edges = cpdb_res |>
    mutate( from=sprintf("%s_%s_ligand", cluster_1, source), to=sprintf("%s_%s_receptor", cluster_2, target)) |>
    select(from, to, name="variable", coef, pvalue, fdr, group, log2_fc) |>
    mutate(log2_fc = if_else(is.infinite(log2_fc) & log2_fc < 0, -max(abs(log2_fc[is.finite(log2_fc)])), log2_fc)) |>
    mutate(log2_fc = if_else(is.infinite(log2_fc) & log2_fc > 0, max(abs(log2_fc[is.finite(log2_fc)])), log2_fc)) |>
    mutate(fdr = if_else(fdr == 0, 1e-255, fdr))

  # make nodes df
  nodes_source = cpdb_res |>
    select(cell_type=cluster_1, gene=source) |>
    mutate(type="ligand", fraction_expressed=0.5) |>
    mutate(name=sprintf("%s_%s_%s", cell_type, gene, type))
  nodes_target = cpdb_res |>
    select(cell_type=cluster_2, gene=target, fraction_expressed) |>
    mutate(type="receptor") |>
    mutate(name=sprintf("%s_%s_%s", cell_type, gene, type))
  nodes = bind_rows(nodes_source, nodes_target) |>
    distinct() |>
    select(name, cell_type, gene, type, fraction_expressed) |>
    filter(name %in% edges$to | name %in% edges$from)

  cell_types = unique(c(cpdb_res$cluster_1, cpdb_res$cluster_2))
  cell_type_lr = unique(paste0(nodes$cell_type, "_", nodes$type))

  # construct hierarchy graph for dendrogram
  edges_ct_hierarchy = bind_rows(
    # root
    data.frame(from="root", to=nodes$cell_type),
    nodes |> mutate(to=paste0(cell_type, "_", type)) |> select(from=cell_type, to) |> distinct(),
    nodes |> mutate(from=paste0(cell_type, "_", type), to=name) |> select(from, to)
  )
  nodes_ct_hierarchy = data.frame(name=unique(c(edges_ct_hierarchy$to, edges_ct_hierarchy$from))) |>
    left_join(nodes, by="name") |> group_by(cell_type) |>  group_modify(function(x, key) {
      n = nrow(x)
      n1 = floor(n/2)
      n2 = n - 1 - n1
      x$cell_type_label = c(rep(NA, n1), key$cell_type, rep(NA, n2))
      x
    }) |> ungroup() |> select(name, cell_type, gene, type, cell_type_label, fraction_expressed)
  G_ct =graph_from_data_frame(edges_ct_hierarchy, directed=TRUE, vertices=nodes_ct_hierarchy)


  # info for geom_conn_bundle
  from = match(edges$from, nodes_ct_hierarchy$name)
  to = match(edges$to, nodes_ct_hierarchy$name)
  conn_meta = data.frame(from=nodes_ct_hierarchy[from,]$name, to=nodes_ct_hierarchy[to,]$name) |>
    left_join(edges, by=c("from", "to"))

  # plot using ggraph
  p = ggraph(G_ct, layout="dendrogram", circular=TRUE) +
    geom_node_point(aes(color=cell_type, shape=type, size=fraction_expressed), stroke=2, fill="white") +
    geom_conn_bundle(
      data = get_con(from=from, to=to, coef=conn_meta$coef, log2_fc=conn_meta$log2_fc, pvalue=conn_meta$pvalue, group=conn_meta$group, fdr=conn_meta$fdr),
      # aes_string(color=edge_color, width="abs(coef)"),
      aes_string(color=edge_color, width="abs(log10(fdr))"),
      tension=0.4,
      arrow=arrow()) +
    edge_color_scale +
    scale_edge_width_continuous(range=c(1,3)) +
    theme_void() +
    geom_text_repel(aes(x=x, y=y, label=gene), size=5 ) +
    geom_label_repel(aes(x=x, y=y, nudge_x=1.3*x, nudge_y=1.3*y, label=cell_type_label), xlim=c(-2, 2), size=8, force=3, alpha=.8) +
    coord_fixed(clip="off") +
    scale_shape_manual(values=c("receptor"=19, "ligand"=21))  +
    scale_color_manual(values=cell_type_color_scale)   +
    scale_size_continuous(range=c(2, 14)) +
    theme(legend.position="top")
  p
}

```

```{r, include=FALSE}
cpdb_res_luad_lusc_coarse = read_csv(file.path(params$input_dir, "cpdb_luad_lusc_coarse.csv"))
```

## Explanation of the plots

 * Each dot in the circle marks a ligand (filled circle) or a receptor (outline square). The direction of the interaction (ligant -> receptor) is additionally indicated with arrows
 * The ligands and receptors are colored and organized by cell-type
 * Arrows indicate interactions that are differentially regulated between two conditions. Red arrows mean up-regulated, blue down-regulated.
 * The thickness of the arrows indicates the confidence (=FDR) that the given interaction is differentially regulated.
 * Usually, the top 100 differentailly regulated interactions are shown. It is, in most cases, not possible to fit all significantly different interactions into the plot.

### LUAD LUSC

Comparison of LUAD and LUSC samples. Blue = Upregulated in LUAD, Red = upregulated in LUSC

```{r}
#
# luad_lusc_interactions = cpdb_res_luad_lusc |>
#   select(source=source_genesymbol, target=target_genesymbol, fraction_expressed, fdr=fdr, log2_fc=log2FoldChange, cluster_2=cell_type_major, pvalue) |>
#   mutate(group = if_else(log2_fc < 0, "LUSC", "LUAD")) |>
#   mutate(cluster_1 = "Tumor cells") |>
#   mutate(variable=sprintf("%s_%s_%s_%s", source, target, cluster_1, cluster_2), coef=log2_fc) |>
#   filter(cluster_2 %in% c(immune_cells))
#
# top_genes = luad_lusc_interactions |> select(source, fdr) |> distinct() |> arrange(fdr) |> pull(source) |> head(10)
```

```{r, fig.width=22, fig.height=14, message=FALSE, warning=FALSE, results=FALSE}
# luad_lusc_interactions  |> filter(source %in% top_genes) |> cpdb_plot(edge_color="log2_fc", cell_type_color_scale = COLORS$cell_type_major)
# ggsave(file.path(params$artifact_dir, "luad_lusc_ciros_cell_type_major.pdf"))
```



```{r}

luad_lusc_interactions_coarse = cpdb_res_luad_lusc_coarse |>
  select(source=source_genesymbol, target=target_genesymbol, fraction_expressed, fdr=fdr, log2_fc=log2FoldChange, cluster_2=cell_type_coarse, pvalue) |>
  mutate(group = if_else(log2_fc < 0, "LUSC", "LUAD")) |>
  mutate(cluster_1 = "Tumor cells") |>
  mutate(variable=sprintf("%s_%s_%s_%s", source, target, cluster_1, cluster_2), coef=log2_fc) |>
  filter(cluster_2 %in% c(immune_cells))

top_genes = luad_lusc_interactions_coarse |> select(source, fdr) |> distinct() |> arrange(fdr) |> pull(source) |> head(10)
```

```{r, fig.width=22, fig.height=14, message=FALSE, warning=FALSE, results=FALSE}
luad_lusc_interactions_coarse  |> filter(source %in% top_genes) |> cpdb_plot(edge_color="log2_fc", cell_type_color_scale = COLORS$cell_type_coarse)
ggsave(file.path(params$artifact_dir, "luad_lusc_ciros_cell_type_coarse.pdf"))
```
